@@ -33,15 +33,15 @@
   },
   "homepage": "https://github.com/CreativeIT/material-angular-dashboard",
   "dependencies": {
-    "@angular/animations": "^7.2.15",
-    "@angular/common": "7.2.1",
-    "@angular/compiler": "7.2.1",
-    "@angular/core": "7.2.1",
-    "@angular/forms": "7.2.1",
+    "@angular/animations": "^8.2.14",
+    "@angular/common": "8.2.14",
+    "@angular/compiler": "8.2.14",
+    "@angular/core": "8.2.14",
+    "@angular/forms": "8.2.14",
     "@angular/http": "7.2.1",
-    "@angular/platform-browser": "7.2.1",
-    "@angular/platform-browser-dynamic": "7.2.1",
-    "@angular/router": "7.2.1",
+    "@angular/platform-browser": "8.2.14",
+    "@angular/platform-browser-dynamic": "8.2.14",
+    "@angular/router": "8.2.14",
     "@types/googlemaps": "^3.30.18",
     "bootstrap": "^4.3.1",
     "classlist.js": "1.1.20150312",
@@ -49,17 +49,17 @@
     "d3": "3.5.17",
     "material-angular-select": "1.0.0",
     "material-design-lite": "1.3.0",
-    "ngx-toastr": "^10.1.0",
-    "node-gyp": "^6.0.0",
+    "ngx-toastr": "^10.2.0",
+    "node-gyp": "^6.0.1",
     "nvd3": "1.8.6",
-    "rxjs": "6.3.3",
-    "zone.js": "0.8.28"
+    "rxjs": "6.5.3",
+    "zone.js": "0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.6.5",
-    "@angular/cli": "7.2.2",
-    "@angular/compiler-cli": "7.2.1",
-    "@angular/language-service": "7.2.1",
+    "@angular-devkit/build-angular": "^0.803.19",
+    "@angular/cli": "8.3.19",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "8.2.14",
     "@types/d3": "3.5.40",
     "@types/material-design-lite": "1.1.15",
     "@types/node": "10.1.2",
@@ -75,6 +75,6 @@
     "tslint": "5.10.0",
     "tslint-angular": "1.1.2",
     "tslint-config-airbnb": "5.9.2",
-    "typescript": "3.2.4"
+    "typescript": "^3.4.5"
   }
 }
